+++ pakyow-assets/lib/pakyow/assets.rb
@@ -5,11 +5,13 @@ require "pakyow/presenter"
 
 require_relative "assets/framework"
 require_relative "behavior/assets/externals"
+require_relative "behavior/assets/multiapp"
 require_relative "behavior/assets/relocate_assets"
 
 module Pakyow
   config.commands.paths << File.expand_path("../commands", __FILE__)
 
   include Behavior::Assets::Externals
+  include Behavior::Assets::Multiapp
   include Behavior::Assets::RelocateAssets
 end