+++ pakyow-assets/lib/pakyow/assets/framework.rb
@@ -10,6 +10,7 @@ module Pakyow
         require "pakyow/application/behavior/assets/packs"
         require "pakyow/application/behavior/assets/silencing"
         require "pakyow/application/behavior/assets/externals"
+        require "pakyow/application/behavior/assets/multiapp"
         require "pakyow/application/behavior/assets/processing"
         require "pakyow/application/behavior/assets/types"
 
@@ -167,6 +168,7 @@ module Pakyow
           include Application::Behavior::Assets::Packs
           include Application::Behavior::Assets::Silencing
           include Application::Behavior::Assets::Externals
+          include Application::Behavior::Assets::Multiapp
           include Application::Behavior::Assets::Processing
           include Application::Behavior::Assets::Types