+++ pakyow-core/spec/features/common/support/project/common/backend/controllers/root.rb
@@ -1,0 +1,9 @@
+controller :root do
+  def self.context
+    :common
+  end
+
+  def self.common?
+    true
+  end
+end